:recycle:Refactor: refactor backtest option states 백테스트 옵션 관련 상태 전역으로 관리 & 코드 분할
