+++ tradin/app/backtest/page.tsx
@@ -1,9 +1,12 @@
 'use client';
 
 import BackTestPage from "../../src/components/units/backtest/main/BackTest";
+import { BackTestProvider } from "../../src/contexts/BackTestContext";
 
 export default function BackTest() {
     return (
-        <BackTestPage />
+        <BackTestProvider>
+            <BackTestPage />
+        </BackTestProvider>
     );
 }
